config: change 'work_mem' and 'maintenance_work_mem' default setting
The current defaults set by this role is pretty high (`work_mem =
512MB`) this is fine on very big machines (with more than 32GB of RAM)
but not compatible with smaller machines (with 4GB of RAM).

This change adds two new optionnal role variables `postgres_work_mem`
and `postgres_maintenance_work_mem` to be able to set a custom size to
your needs.

It also tries to set sensible defaults to:
- default value for `work_mem` to 0.5 % of the total available memory
  of the machine
- default value for `maintenance_work_mem` to 5% of the total
available memory of the machine

P.S.: here is a good explanation about those configuration option
https://info.crunchydata.com/blog/optimize-postgresql-server-performance