fix: use modern way of passing variables to included task file
Since a while (I believe 2.6) ansible doesn't allow variables
passing with the `include:` statement. We need to explicit the `vars:`
argument for it to work properly.

I have also changed the `with_subelements` loop to use the more modern
`loop` iterator.

This fixes a bug where users / databases were not created on pg
clusters when you are trying to configure multiple clusters on a
host (when `postgres_clusters|length > 1`).